Condensed the Action files
