--- UI/libs/shared/ui/src/lib/helpers/storeHelpers/DOBActions.ts
@@ -1,6 +1,0 @@
-import { store } from '@shared-ui/store';
-import { DOB } from '@shared-ui/types/defualtTypes';
-
-export async function addDOBAction(dob: DOB) {
-  await store.dispatch('addDOB', dob);
-}
