--- UI/libs/shared/ui/src/lib/helpers/storeHelpers/aliasActions.ts
@@ -1,5 +1,0 @@
-import { store } from '@shared-ui/store';
-
-export async function addAliasesAction(aliases) {
-  await store.dispatch('addAlias', aliases);
-}
