--- UI/libs/shared/ui/src/lib/helpers/storeHelpers/citizenActions.ts
@@ -1,6 +1,0 @@
-import { store } from '@shared-ui/store';
-import { Citizenship } from '@shared-ui/types/defualtTypes';
-
-export async function addCitizenshipAction(citizenship: Citizenship) {
-  await store.dispatch('addCitizenshipInfo', citizenship);
-}
