--- UI/libs/shared/ui/src/lib/helpers/storeHelpers/idActions.ts
@@ -1,6 +1,0 @@
-import { store } from '@shared-ui/store';
-import { Id } from '@shared-ui/types/defualtTypes';
-
-export async function addIdAction(id: Id) {
-  await store.dispatch('addId', id);
-}
