--- UI/libs/shared/ui/src/lib/helpers/storeHelpers/personalInfoActions.ts
@@ -1,5 +1,0 @@
-import { store } from '@shared-ui/store';
-
-export async function addPersonalInfoAction(personalInfo) {
-  store.dispatch('addPersonalInfo', personalInfo);
-}
